fix bugs

+++ pack_labs_02/lab_1/include/fraction.h
@@ -7,8 +7,8 @@
 
 
 
-int* reducing_a_fraction(const int numerator, const int denominator, error_code_t* error);
-void numerator_and_denominator_from_fraction(const double number, int*numerator, int*denominator, error_code_t* error);
+long long* reducing_a_fraction(const long long numerator, const long long denominator, error_code_t* error);
+void numerator_and_denominator_from_fraction(const double number, long long*numerator, long long*denominator, error_code_t* error);
 
 
 #endif
